Simplify one-to-many relationships by removing parent refs
Before, domain models had reference properties to their owners. For example, the column type had a property
`projectId` which was the id of the project owning the column. After deliberation and some fighting with
our database model, we have decided to use a simplified approach where only the parent will have a list
of the children in one-to-many relationships.