Update create project feature to require authentication as per #29
The `createProject` command handler has been updated to require a valid user id to execute. This shall only
be passed in when a user is authenticated.

A number of changes have been made to update the database models as per #43. Many files were touched to
modify every layer involved.

One such modification was to return null entities when an entity cannot be found, as per #44. In addition,
unique errors such as `NotFoundError` are now throw by the persistance layer instead of throwing generic
`Error`s.