Delete get project columns feature to reflect embedded design as per #43
Since we are now using am embedded design, we don't need to query for columns separately; they
are included as embedded documents when a project is queried for. Therefore, we don't need a separate
feature to get the columns of a feature.

This also closes #33 (which is outdated at this point because of #43)