Update column and issue db models to have permanent id properties
When a project is updated, the subdocuments (columns and issues) sometimes refresh
their `_id` property, which in turn changes the `id`. We need these ids to be permanent, so
we added a `columnId` and `issueId` property, respectively, to each model. These properties are generated
once when the document is created and never changed thereafter.